Курсовой проект без комментариев.

+++ src/main/java/com/example/auction/services/LotService.java
@@ -5,7 +5,6 @@ import com.example.auction.enums.LotStatus;
 import com.example.auction.models.Lot;
 import com.example.auction.repositories.BidRepository;
 import com.example.auction.repositories.LotRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
@@ -19,13 +18,8 @@ import java.util.stream.Collectors;
 public class LotService {
     private LotRepository lotRepository;
     private BidRepository bidRepository;
-//    private BidDTO bidDTO;
     private BidDTOForFullLotDTO bidDTOForFullLotDTO;
-//    public LotService(LotRepository lotRepository) {
-//        this.lotRepository = lotRepository;
-//    }
 
-//    @Autowired
     public LotService(BidRepository bidRepository, LotRepository lotRepository) {
         this.bidRepository = bidRepository;
         this.lotRepository = lotRepository;
@@ -50,7 +44,7 @@ public class LotService {
                 .map(FullLotDTO::fromLotToFullLotDTO)
                 .collect(Collectors.toList());
     }
-    private Long totalPrice(Long lotId, Integer bidPrice, Integer startPrice) {
+    private Long totalPrice(Long lotId, Long bidPrice, Long startPrice) {
         return (bidRepository.getCountNumberOfBidByLotId(lotId) * bidPrice + startPrice);
     }
     public FullLotDTO getFullLotById (Long id) {
@@ -66,10 +60,6 @@ public class LotService {
         bidDTOForFullLotDTO.setBidDate(bidRepository.findByBidDateMax(lotId).getBidDate());
         return bidDTOForFullLotDTO;
     }
-//    public LotDTO updateStatusStopped (Long lotId, String lotStatus) {
-//        Lot updateInfoAboutLot = lotRepository.updateLotStatus(lotId, lotStatus);
-//        return LotDTO.fromLot(updateInfoAboutLot);
-//    }
     public LotDTO getLotById (Long id) {
         return LotDTO.fromLotToLotDTO(lotRepository.findById(id).get());
     }
@@ -83,66 +73,6 @@ public class LotService {
         lot.setStatus(LotStatus.STOPPED);
         lotRepository.save(lot);
     }
-    //    public Collection<LotDTO> getLotsByStatus(String status){
-//        return lotRepository.findByStatus(status)
-//                .stream()
-//                .map(LotDTO::fromLot)
-//                .collect(Collectors.toList());
-//    }
-    //    public LotDTO updateStatus (Long id, String status) {
-//        Lot updateInfoAboutLot = lotRepository.updateLotStatus(id, status);
-//        return LotDTO.fromLot(updateInfoAboutLot);
-//    }
-//    public String findStatus (Long id) {
-//        String lotStatus = String.valueOf(lotRepository.findLotStatusByLotId(id));
-//        return lotStatus;
-//    }
-//    public LotDTO updateInfoAboutLot (LotDTO lotDTO) {
-//        Lot lot = lotDTO.toLot();
-//        Lot updateInfoAboutLot = lotRepository.save(lot);
-//        return LotDTO.fromLot(updateInfoAboutLot);
-//    }
-//
-//    public void deleteLotById (Long lotId) {
-//        lotRepository.deleteById(lotId);
-//    }
-//    public List<Lot> getAllLotsByPage (Integer pageNumber, Integer pageSize, String lotStatus) {
-//        PageRequest pageRequest = PageRequest.of(pageNumber - 1, pageSize);
-//        return lotRepository.findAll(pageRequest).getContent();
-//    }
-//
-//    public Collection<LotDTO> getAllLots() {
-//        return lotRepository.findAll()
-//                .stream()
-//                .map(LotDTO::fromLot)
-//                .collect(Collectors.toList());
-//    }
-//
-//    public Collection<BidDTO> getBiddersByLotId(Long lotId) {
-//        List<Bid> bids = lotRepository.findById(lotId).get().getBids();
-//        List<BidDTO> bidDTOs = new ArrayList<>();
-//        for(Bid bid : bids) {
-//            BidDTO bidDTO = BidDTO.fromBid(bid);
-//            bidDTOs.add(bidDTO);
-//        }
-//        return bidDTOs;
-//    }
-//    public LotDTO getLotByTitle (String lotTitle) {
-//        return LotDTO.fromLot(lotRepository.findByLotTitleContainsIgnoreCase(lotTitle));
-//    }
-//    public Collection<FullLotDTO> getAllLotsForExport(Long bidderId, Long lotId) {
-//        return lotRepository.findAll().stream()
-//                .map(FullLotDTO::fromLot)
-//                .peek(lot -> (lot).setLotCurrentPrice(sumCurrentPrice(lot.getLotId(), lot.getLotLastBid(), lot.getLotStartPrice())))
-//                .peek(lot -> (lot).setLotBidPrice(findInfoABoutLastBid(bidderId,lotId)
-//                .getLotBidPrice();
-//    }
-
-    //    public FullLotDTO createNewFullLot (FullLotDTO fullLotDTO) {
-//        Lot lot = fullLotDTO.toFullLot();
-//        Lot createNewFullLot = lotRepository.save(lot);
-//        return FullLotDTO.fromFullLot(createNewFullLot);
-//    }
     public boolean checkMistakeInCreatingLot(CreateLotDTO createLotDTO) {
         if(createLotDTO.getTitle() == null || createLotDTO.getTitle().isBlank() ||
                 createLotDTO.getDescription() == null || createLotDTO.getDescription().isBlank() ||
